Removed edit functionality, got delete functionality to work.

+++ public/js/playlistDelete.js
@@ -1,14 +1,20 @@
 const delButtonHandler = async (event) => {
   if (event.target.hasAttribute('data-id')) {
     const id = event.target.getAttribute('data-id');
-
+    console.log(event.target);
+    let killSummer = confirm(
+      'Are you sure you want to delete me?\n\nThis procdure is irreversible...'
+    );
+    if (!killSummer) {
+      return;
+    }
     try {
       const response = await fetch(`/api/playlist/${id}`, {
         method: 'DELETE',
       });
 
       if (response.ok) {
-        document.location.replace('/user');
+        document.location.replace('/');
       } else {
         const responseData = await response.json();
         alert(responseData.message || 'Failed to delete playlist');
